+++ src/config/mod.rs
@@ -1,3 +1,4 @@
+pub mod filters;
 pub mod structs;
 
 use std::{