+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod add;
 pub mod config;
 pub mod file_picker;
+pub mod iter_ext;
 pub mod modpack;
 pub mod scan;
 pub mod upgrade;